[Refactor] Split Elasticsearch connector into dedicated backend library
