+++ be/CMakeLists.txt
@@ -193,6 +193,7 @@ option(WITH_BRCP_KEEPALIVE "Build binary with brpc_socket_keepalive (requires br
 
 option(WITH_BENCH "Build binary with bench" OFF)
 option(WITH_CONNECTOR_BENCHMARK "Build the benchgen-backed benchmark connector" ON)
+option(WITH_CONNECTOR_ELASTICSEARCH "Build the Elasticsearch connector" ON)
 
 option(USE_STAROS "Use StarOS to manager tablet info" OFF)
