+++ be/src/connector/connector_registry_init.cpp
@@ -18,7 +18,6 @@
 #include "connector/builtin_connector_registry.h"
 #include "connector/cache_stats_connector.h"
 #include "connector/connector_registry.h"
-#include "connector/es_connector.h"
 #include "connector/file_connector.h"
 #include "connector/hive_connector.h"
 #ifndef __APPLE__
@@ -44,7 +43,6 @@ void install_if_absent(ConnectorRegistry* registry, const std::string& name) {
 Status install_builtin_connectors(ConnectorRegistry* registry) {
     DCHECK(registry != nullptr);
     install_if_absent<HiveConnector>(registry, Connector::HIVE);
-    install_if_absent<ESConnector>(registry, Connector::ES);
     install_if_absent<JDBCConnector>(registry, Connector::JDBC);
     install_if_absent<MySQLConnector>(registry, Connector::MYSQL);
     install_if_absent<CacheStatsConnector>(registry, Connector::CACHE_STATS);
