+++ be/src/connector/elasticsearch/es_scan_reader.cpp
@@ -32,7 +32,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-#include "exec/es/es_scan_reader.h"
+#include "connector/elasticsearch/es_scan_reader.h"
 
 #include <map>
 #include <sstream>
@@ -42,8 +42,8 @@
 #include "common/logging.h"
 #include "common/status.h"
 #include "common/thread/priority_thread_pool.hpp"
-#include "exec/es/es_scroll_parser.h"
-#include "exec/es/es_scroll_query.h"
+#include "connector/elasticsearch/es_scroll_parser.h"
+#include "connector/elasticsearch/es_scroll_query.h"
 #include "fmt/compile.h"
 
 namespace starrocks {
