+++ build.sh
@@ -116,6 +116,8 @@ Usage: $0 <options>
      --with-bench       build Backend with bench(default without bench)
      --without-connector-benchmark
                         build Backend without the benchgen-backed benchmark connector
+     --without-connector-elasticsearch
+                        build Backend without the Elasticsearch connector
      --with-dynamic     build Backend with dynamic linking of individual StarRocks modules (developer option)
      --with-clang-tidy  build Backend with clang-tidy(default without clang-tidy)
      --without-java-ext build Backend without java-extensions(default with java-extensions)
@@ -172,6 +174,7 @@ OPTS=$(${GETOPT_BIN} \
   -l 'with-gcov' \
   -l 'with-bench' \
   -l 'without-connector-benchmark' \
+  -l 'without-connector-elasticsearch' \
   -l 'with-dynamic' \
   -l 'module' \
   -l 'with-clang-tidy' \
@@ -210,6 +213,7 @@ CLEAN=
 WITH_GCOV=OFF
 WITH_BENCH=OFF
 WITH_CONNECTOR_BENCHMARK=ON
+WITH_CONNECTOR_ELASTICSEARCH=ON
 WITH_CLANG_TIDY=OFF
 WITH_COMPRESS=ON
 THIN_ARCHIVE=OFF
@@ -337,6 +341,7 @@ else
             --enable-shared-data|--use-staros) USE_STAROS=ON; shift ;;
             --with-bench) WITH_BENCH=ON; shift ;;
             --without-connector-benchmark) WITH_CONNECTOR_BENCHMARK=OFF; shift ;;
+            --without-connector-elasticsearch) WITH_CONNECTOR_ELASTICSEARCH=OFF; shift ;;
             --with-dynamic) ENABLE_MULTI_DYNAMIC_LIBS=ON; shift ;;
             --module) BUILD_BE_MODULE=$2; shift 2 ;;
             --with-clang-tidy) WITH_CLANG_TIDY=ON; shift ;;
@@ -401,6 +406,7 @@ echo "Get params:
     WITH_GCOV                   -- $WITH_GCOV
     WITH_BENCH                  -- $WITH_BENCH
     WITH_CONNECTOR_BENCHMARK    -- $WITH_CONNECTOR_BENCHMARK
+    WITH_CONNECTOR_ELASTICSEARCH -- $WITH_CONNECTOR_ELASTICSEARCH
     WITH_CLANG_TIDY             -- $WITH_CLANG_TIDY
     WITH_COMPRESS_DEBUG_SYMBOL  -- $WITH_COMPRESS
     THIN_ARCHIVE                -- $THIN_ARCHIVE
@@ -552,6 +558,7 @@ if [ ${BUILD_BE} -eq 1 ] || [ ${BUILD_FORMAT_LIB} -eq 1 ] ; then
                   -DENABLE_QUERY_DEBUG_TRACE=$ENABLE_QUERY_DEBUG_TRACE  \
                   -DWITH_BENCH=${WITH_BENCH}                            \
                   -DWITH_CONNECTOR_BENCHMARK=${WITH_CONNECTOR_BENCHMARK} \
+                  -DWITH_CONNECTOR_ELASTICSEARCH=${WITH_CONNECTOR_ELASTICSEARCH} \
                   -DENABLE_MULTI_DYNAMIC_LIBS=${ENABLE_MULTI_DYNAMIC_LIBS}\
                   -DWITH_CLANG_TIDY=${WITH_CLANG_TIDY}                  \
                   -DWITH_COMPRESS=${WITH_COMPRESS}                      \
